+++ .gitignore
@@ -167,3 +167,5 @@ cython_debug/
 # predictoor dynamic modeling
 out*.txt
 csvs/
+
+pdr_backend/accuracy/output/*.json
