+++ pdr_backend/accuracy/app.py
@@ -1,9 +1,8 @@
 from flask import Flask, jsonify, request, abort
 import threading
 import json
-from utils.extract_statistics import extract_statistics
-from utils.get_all_predictions import get_all_predictions
-from utils.get_all_contracts import get_all_contracts
+from pdr_backend.util.predictoor_stats import get_endpoint_statistics
+from pdr_backend.util.subgraph_predictions import get_all_predictions, get_all_contracts, FilterMode
 from utils.get_start_end_params import get_start_end_params
 
 app = Flask(__name__)
@@ -16,8 +15,8 @@ def save_predictions_to_file():
 
             start_ts_param, end_ts_param = get_start_end_params()
             contract_addresses = get_all_contracts("0x4ac2e51f9b1b0ca9e000dfe6032b24639b172703", network_param)
-            predictions = get_all_predictions(start_ts_param, end_ts_param, contract_addresses, network_param)
-            statistics = extract_statistics(predictions)
+            predictions = get_all_predictions(start_ts_param, end_ts_param, contract_addresses, network_param, filter_mode=FilterMode.CONTRACT)
+            statistics = get_endpoint_statistics(predictions)
 
             with open(JSON_FILE_PATH, 'w') as f:
                 json.dump({
@@ -32,6 +31,7 @@ def save_predictions_to_file():
 
 @app.route('/predictions', methods=['GET'])
 def serve_predictions_from_file():
+    print("Serving predictions data from file")
     try:
         with open(JSON_FILE_PATH, 'r') as f:
             data = json.load(f)
