+++ pdr_backend/util/predictoor_stats.py
@@ -1,16 +1,15 @@
 from typing import List, Dict, Tuple, Union
-from models.prediction import Prediction
+from pdr_backend.models.prediction import Prediction
 
 PredictionDetail = Tuple[str, str, str]
 PredictoorStats = Dict[str, Union[str, float, int, List[PredictionDetail]]]
 PairTimeframeStats = Dict[str, Union[str, float, int]]
 StatisticsResult = Dict[str, Union[float, List[PredictoorStats], List[PairTimeframeStats]]]
 
-def extract_statistics(all_predictions: List[Prediction]) -> StatisticsResult:
-    # This is similar to the get_statistics function but returns the results instead of printing them.
-    total_predictions = len(all_predictions)
-    correct_predictions = 0
+
+def get_pure_stats(all_predictions: List[Prediction]):
     stats = {"pair_timeframe": {}, "predictoor": {}}
+    correct_predictions = 0
 
     for prediction in all_predictions:
         pair_timeframe_key = (prediction.pair, prediction.timeframe)
@@ -52,6 +51,14 @@ def extract_statistics(all_predictions: List[Prediction]) -> StatisticsResult:
             (prediction.pair, prediction.timeframe, source)
         )
 
+    return stats, correct_predictions
+
+
+def get_endpoint_statistics(all_predictions: List[Prediction]) -> StatisticsResult:
+    # This is similar to the get_statistics function but returns the results instead of printing them.
+    total_predictions = len(all_predictions)
+    stats, correct_predictions = get_pure_stats(all_predictions)
+
     overall_accuracy = correct_predictions / total_predictions * 100 if total_predictions else 0
 
     results = {
@@ -84,3 +91,29 @@ def extract_statistics(all_predictions: List[Prediction]) -> StatisticsResult:
         })
 
     return results
+
+
+def get_cli_statistics(all_predictions):
+    total_predictions = len(all_predictions)
+    
+    stats, correct_predictions = get_pure_stats(all_predictions)
+    print(f"Overall Accuracy: {correct_predictions/total_predictions*100:.2f}%")
+
+    for key, data in stats["pair_timeframe"].items():
+        pair, timeframe = key
+        accuracy = data["correct"] / data["total"] * 100
+        print(f"Accuracy for Pair: {pair}, Timeframe: {timeframe}: {accuracy:.2f}%")
+        print(f"Total stake: {data['stake']}")
+        print(f"Total payout: {data['payout']}")
+        print(f"Number of predictions: {data['total']}\n")
+
+    for predictoor, data in stats["predictoor"].items():
+        accuracy = data["correct"] / data["total"] * 100
+        print(f"Accuracy for Predictoor Address: {predictoor}: {accuracy:.2f}%")
+        print(f"Stake: {data['stake']}")
+        print(f"Payout: {data['payout']}")
+        print(f"Number of predictions: {data['total']}")
+        print("Details of Predictions:")
+        for detail in data["details"]:
+            print(f"Pair: {detail[0]}, Timeframe: {detail[1]}, Source: {detail[2]}")
+        print("\n")
